Move refactored fns into new (poorly named) file
