+++ cpp/purify/CMakeLists.txt
@@ -35,7 +35,7 @@ set(HEADERS
 set(SOURCES utilities.cc pfitsio.cc
   kernels.cc wproj_utilities.cc operators.cc uvfits.cc yaml-parser.cc
   read_measurements.cc distribute.cc integration.cc wide_field_utilities.cc wkernel_integration.cc
-  wproj_operators.cc uvw_utilities.cc)
+  wproj_operators.cc uvw_utilities.cc setup_utils.cc)
 
 if(TARGET casacore::ms)
   list(APPEND SOURCES casacore.cc)
